+++ scripts/module.mjs
@@ -1,6 +1,7 @@
 import { initSettings } from "./settings.mjs"
 import { initQueries } from "./queries.mjs";
 import { readySaveTray } from "./message-handler.mjs"
+import { readyTemplateTargeting } from "./template-targeting.mjs";
 
 Hooks.once("init", () => {
   initSettings();
@@ -9,4 +10,5 @@ Hooks.once("init", () => {
 
 Hooks.once("ready", () => {
   readySaveTray();
+  readyTemplateTargeting();
 });